requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.3.0
Automat==20.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.4
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==3.3.1
cssselect==1.1.0
distlib==0.3.1
filelock==3.0.12
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==20.0.1
ibm-cloud-sdk-core==3.5.2
ibm-watson==5.1.0
identify==1.5.13
idna==2.10
incremental==17.5.0
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
joblib==1.0.0
lxml==4.6.2
MouseInfo==0.1.3
multidict==5.1.0
mutagen==1.45.1
nltk==3.5
nodeenv==1.5.0
numpy==1.19.5
packaging==20.9
pandas==1.2.3
parsel==1.6.0
Pillow==8.1.2
pluggy==0.13.1
pre-commit==2.10.0
Protego==0.1.16
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycparser==2.20
PyDispatcher==2.0.5
PyGetWindow==0.0.9
PyHamcrest==2.0.2
PyJWT==2.0.1
PyMsgBox==1.0.9
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==2.8.1
python3-xlib==0.15
PyTweening==1.0.3
pytz==2021.1
PyYAML==5.4.1
queuelib==1.5.0
regex==2020.11.13
requests==2.25.1
rfc3986==1.4.0
Scrapy==2.4.1
service-identity==18.1.0
six==1.15.0
sniffio==1.2.0
soupsieve==2.2
SpeechRecognition==3.8.1
toml==0.10.2
tox==3.21.3
tqdm==4.56.0
translate==3.5.0
Twisted==20.3.0
typing-extensions==3.7.4.3
urllib3==1.26.2
vaderSentiment==3.3.2
virtualenv==20.4.2
w3lib==1.22.0
webencodings==0.5.1
websocket-client==0.48.0
yarl==1.6.3
zope.interface==5.2.0
